examples/image_classification/mnist/requirements.txt
pandas>=0.23.4
keras>=2.2.4
tensorflow-gpu>=1.11.0
numpy>=1.15.2
scipy>=1.1.0
scikit-learn>=0.20.0
scikit-image>=0.14.1
alectio_sdk==0.0.1
Pillow==7.1.2
tqdm==4.46.0
urllib3==1.25.9
wasabi==0.6.0
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
flask
pyyaml
tensorflow-datasets
